Move url-input-component to shared
