+++ apps/developer-portal/src/components/shared/index.ts
@@ -23,3 +23,4 @@ export * from "./empty-placeholder";
 export * from "./avatar-svg-background";
 export * from "./upload-file";
 export * from "./dynamic-fields";
+export * from "./url-input-component";
